fix:16byteになるようにheaderを調整

+++ robot_port/include/robot_port/robot_data_config.hpp
@@ -20,7 +20,7 @@ struct feedback_data_t {
 } __attribute__((__packed__));
 
 struct operation_data_t {
-    uint16_t header;  // ヘッダー
+    uint32_t header;  // ヘッダー
     float vx;         // x軸方向の速度[m/s]
     float vy;         // y軸方向の速度[m/s]
     float omega;      // 回転速度[rad/s]
